Split fr and fp classes out of odin_data.py
